requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.6
joblib==1.3.2
multidict==6.0.4
numpy==1.26.3
python-dotenv==1.0.1
regex==2023.12.25
requests==2.31.0
scipy==1.12.0
threadpoolctl==3.2.0
tiktoken==0.5.2
urllib3==2.1.0
yarl==1.9.4
